Combined load-common and load-context
