+++ src/backend/load/load-group.ts
@@ -20,7 +20,7 @@ import {
     type InsertableTarget,
     type LoadContext,
     getOnshapeApiFromContext
-} from "./load-context";
+} from "./load-common";
 import { uploadThumbnailsStep } from "./load-steps";
 import type { InstancePath } from "../../shared/onshape-path";
