+++ src/backend/load/workflows.ts
@@ -19,9 +19,10 @@ import { group, libraries } from "../../shared/schema";
 import {
     type GroupTarget,
     type LoadContext,
+    LOAD_CONCURRENCY,
+    createLimiter,
     getOnshapeApiFromContext
-} from "./load-context";
-import { LOAD_CONCURRENCY, createLimiter } from "./load-common";
+} from "./load-common";
 import { untrackJob } from "./job-tracker";
 import { loadGroup } from "./load-group";
