+++ src/backend/parse/parse-part-number.ts
@@ -34,7 +34,7 @@ import type {
 import {
     type LoadContext,
     getOnshapeApiFromContext
-} from "../load/load-context";
+} from "../load/load-common";
 import { ONSHAPE_STEP_RETRIES } from "../load/load-steps";
 
 /** Configurations fetched per workflow step. */
