+++ src/frontend/groups/add-group-menu.tsx
@@ -50,7 +50,7 @@ function AddGroupMenuContent(props: AddGroupMenuContentProps): ReactNode {
             // Replace the loading toast with a transient "started" toast; the
             // spinner signals the running load and the navbar watcher reports
             // completion.
-            showInfoToast("Adding document…", "add-group");
+            showInfoToast("Adding document...", "add-group");
             void queryClient.invalidateQueries({
                 queryKey: jobStatusQueryKey(libraryId)
             });
